# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "server API for Holy Hospital Application.",
        "title": "Holy Hospital Sever API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ec2-3-91-177-221.compute-1.amazonaws.com",
    "basePath": "/api",
    "paths": {
...
        "/outpatient": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Fetch All Patient with New Medic Record Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "GetAllPatient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date Filter Range Start (2022-02-23)",
                        "name": "date_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date Filter Range End (2006-02-25)",
                        "name": "date_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.OutPatient"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Add New Medic Record for Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "New Medic Record",
                "parameters": [
                    {
                        "description": "New Medic Record",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AdminMedicRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/outpatient/:id": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Process Medic Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "Process Doctor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.OutPatient"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/outpatient/:id/assign_nurse": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Process Medic Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "Process Doctor",
                "parameters": [
                    {
                        "description": "Assign Nurse to Medical Check",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AssignNurse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/outpatient/:id/process": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Process Medic Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "Process Doctor",
                "parameters": [
                    {
                        "description": "Process Medic Record by Doctor",
                        "name": "body_doctor",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.DoctorMedicRecord"
                        }
                    },
                    {
                        "description": "Process Medic Record by Nurse",
                        "name": "body_nurse",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.NurseMedicRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageOnly"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/outpatient/log": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Show Report of Submitted data by Medical Staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "Report Log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.OutPatientReportLog"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/outpatient/report": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "Show Report of Submitted data by All Medical Staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OutPatient"
                ],
                "summary": "Report Log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.MessageData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.OutPatientReportLog"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AdminMedicRecord": {
            "type": "object",
            "properties": {
                "complaint": {
                    "type": "string",
                    "example": "sakit perut"
                },
                "date_check": {
                    "type": "string",
                    "example": "2022-06-24"
                },
                "medical_facility_id": {
                    "type": "integer",
                    "example": 1
                },
                "medical_staff_id": {
                    "type": "integer",
                    "example": 1
                },
                "patient_id": {
                    "type": "integer",
                    "example": 1
                },
                "session_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "request.AssignNurse": {
            "type": "object",
            "properties": {
                "nurse_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "request.DoctorMedicRecord": {
            "type": "object",
            "properties": {
                "diagnose": {
                    "type": "string",
                    "example": "maag"
                },
                "prescription": {
                    "type": "string",
                    "example": "entrostop"
                }
            }
        },
...
        "request.NurseMedicRecord": {
            "type": "object",
            "properties": {
                "blood_tension": {
                    "type": "integer",
                    "example": 122
                },
                "body_temp": {
                    "type": "integer",
                    "example": 31
                },
                "height": {
                    "type": "integer",
                    "example": 55
                },
                "weight": {
                    "type": "integer",
                    "example": 165
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.MessageOnly": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.OutPatient": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "complaint": {
                    "type": "string"
                },
                "date_check": {
                    "type": "string"
                },
                "doctor": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "queue": {
                    "type": "integer"
                },
                "serial_number": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                }
            }
        },
        "response.OutPatientReportLog": {
            "type": "object",
            "properties": {
                "serial_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}